create callable class for metric
