+++ otito/metrics/_base_metric.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class BaseMetric(ABC):
+    @abstractmethod
+    def calculate(self, *args, **kwargs):
+        pass
