+++ otito/metrics/numpy/__init__.py
@@ -1,0 +1,3 @@
+from otito.metrics.numpy.classification.accuracy import Accuracy
+
+__all__ = ["Accuracy"]
